+++ src/conn.h
@@ -1,5 +1,5 @@
 #include "common.h"
 #include "libvici.h"
 
-void add_list(vici_req_t* req, char* type, char* proposals[], uint64_t len);
+void add_list(vici_req_t* req, char* type, char* proposals[], uint32_t len);
 int setup(char* remote);
